chore(audit): remove stale runs.rs baseline rows after split
