+++ homeboy.json
@@ -668,8 +668,8 @@
         "field_patterns::src/commands/agent_task/args/controller.rs::RepeatedFieldPattern",
         "field_patterns::src/commands/fuzz.rs::RepeatedFieldPattern",
         "field_patterns::src/commands/report/performance_digest.rs::RepeatedFieldPattern",
-        "field_patterns::src/commands/runs.rs::RepeatedFieldPattern",
         "field_patterns::src/commands/runs/gh_actions.rs::RepeatedFieldPattern",
+        "field_patterns::src/commands/runs/types.rs::RepeatedFieldPattern",
         "field_patterns::src/commands/trace/output.rs::RepeatedFieldPattern",
         "field_patterns::src/core/agent_task_contract.rs::RepeatedFieldPattern",
         "field_patterns::src/core/agent_task_fanout.rs::RepeatedFieldPattern",
@@ -748,7 +748,6 @@
         "structural::src/commands/runner.rs::GodFile",
         "structural::src/commands/runner.rs::HighItemCount",
         "structural::src/commands/runner/doctor.rs::GodFile",
-        "structural::src/commands/runs.rs::GodFile",
         "structural::src/commands/runs/gh_actions.rs::HighItemCount",
         "structural::src/commands::DirectorySprawl",
         "structural::src/core/agent_task.rs::GodFile",
